Created slider menu item behaviour.
Created a new script, SliderBehaviour that defines behaviour for slider type menu items: slider has minimum and maximum values and amount of step. When the user moves mouse to left on the screen, slider moves towards minimum value. When the user moves mouse to the right, the slider moves towards maximum value.